+++ src/api/signin.ts
@@ -1,6 +1,6 @@
 import { AxiosError, AxiosResponse } from "axios";
 import { SIGNIN_ENDPOINT } from "../constants/constants";
-import { client } from "./client";
+import { authClient } from "./client";
 
 /**
  * @param {string} email
@@ -9,10 +9,9 @@ import { client } from "./client";
  */
 async function signin(email: string, password: string) {
   try {
-    const res: AxiosResponse<{ access_token: string }> = await client.post(
+    const res: AxiosResponse<{ access_token: string }> = await authClient.post(
       SIGNIN_ENDPOINT,
-      { email, password },
-      { headers: { "Content-Type": "application/json" } }
+      { email, password }
     );
     if (res.status === 200) return res.data;
   } catch (error) {
